app/serves/fetch_web_resouece/requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.3
diskcache==5.6.3
extractors==66.0.3
fastapi==0.115.0
htmldate==1.9.0
lxml==5.3.0
orjson==3.10.7
primp==0.6.3
PyAutoGUI==0.9.33
pydantic==2.9.2
python-dotenv==1.0.1
Requests==2.32.3
utils==1.0.2
uvicorn==0.31.1
validators==0.34.0
